backend/requirements.txt
# Django
Django>=5.0.0,<6.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5
djangorestframework-simplejwt>=5.3.0

# PostgreSQL
psycopg2-binary>=2.9.9

# Environment
python-decouple>=3.8
python-dotenv>=1.0.0

# Utilities
Pillow>=10.2.0
pytz>=2024.1

# Development
black>=24.1.0
flake8>=7.0.0
pytest>=7.4.4
pytest-django>=4.7.0
factory-boy>=3.3.0
